requirements.txt
playwright>=1.40
python-dotenv>=1.0
pymupdf>=1.24
openai>=1.30
msal>=1.24
requests>=2.31
pytest>=8.0
